Whwn the HackRF changes mode, its parameters have to be changed, e.g if the TX frequency is different from the RX frequency. These parameters are already locally held in the _tx_stream and _rx_stream structs and are already used when shifting between receiving and transmitting. However the code that accepts frequency changes always leaked these changes into the HackRF, even if the frequency change was not meant for the current mode the HackRF was in. Thus, a change of only the transmit frequency affected the actual  receiving frequency of the HackRF when in receiving mode and  vice versa when the receive frequency was changed when   in transmit mode.
This change fixes this problem by only changing the HackRF actual frequency when the direction given to setFrequency is the same as the HackRF current mode.

When the mode changes, the HackRF frequency will be changed to the
current frequency stored in the stream struct, but this is
something that was already implemented.